fixed decimal input for price input and display of next PO numbers in PO forms

+++ home/forms.py
@@ -32,7 +32,7 @@ class purchase_order_form(forms.Form):
 
 	item_1_description = forms.CharField(required=False, widget = forms.TextInput(attrs={'id':'item_1_description_input'}))
 	item_1_fullname = forms.CharField(required=False, widget = forms.TextInput(attrs={'id':'item_1_fullname_input'}))
-	item_1_price = forms.IntegerField(required=False, widget = forms.NumberInput(attrs={'id':'item_1_price_input'}))
+	item_1_price = forms.DecimalField(required=False, widget = forms.NumberInput(attrs={'id':'item_1_price_input'}))
 	item_1_job = forms.ModelChoiceField(queryset=Jobs.objects.all(), to_field_name="address", required=False, widget = forms.Select(attrs={'id':'item_1_job_input'})) # CHANGE LATER TO ONLY ACTIVE JOBS ON THIS
 	item_1_delivery_location = forms.ChoiceField(choices=(('shop', 'shop'),('site', 'site')), required=False, widget= forms.Select(attrs={'id':'item_1_delivery_location_input'})) # make it a radio widget, see docs on widgets
 	item_1_delivery_date = forms.DateField(required=False, widget = forms.SelectDateWidget(attrs={'id':'item_1_delivery_date_input'}))
@@ -40,7 +40,7 @@ class purchase_order_form(forms.Form):
 
 	item_2_description = forms.CharField(required=False, widget = forms.TextInput(attrs={'id':'item_2_description_input'}))
 	item_2_fullname = forms.CharField(required=False, widget = forms.TextInput(attrs={'id':'item_2_fullname_input'}))
-	item_2_price = forms.IntegerField(required=False, widget = forms.NumberInput(attrs={'id':'item_2_price_input'}))
+	item_2_price = forms.DecimalField(required=False, widget = forms.NumberInput(attrs={'id':'item_2_price_input'}))
 	item_2_job = forms.ModelChoiceField(queryset=Jobs.objects.all(), to_field_name="address", required=False, widget = forms.Select(attrs={'id':'item_2_job_input'})) # CHANGE LATER TO ONLY ACTIVE JOBS ON THIS
 	item_2_delivery_location = forms.ChoiceField(choices=(('shop', 'shop'),('site', 'site')), required=False, widget= forms.Select(attrs={'id':'item_2_delivery_location_input'})) # make it a radio widget, see docs on widgets
 	item_2_delivery_date = forms.DateField(required=False, widget = forms.SelectDateWidget(attrs={'id':'item_2_delivery_date_input'}))
@@ -48,7 +48,7 @@ class purchase_order_form(forms.Form):
 
 	item_3_description = forms.CharField(required=False, widget = forms.TextInput(attrs={'id':'item_3_description_input'}))
 	item_3_fullname = forms.CharField(required=False, widget = forms.TextInput(attrs={'id':'item_3_fullname_input'}))
-	item_3_price = forms.IntegerField(required=False, widget = forms.NumberInput(attrs={'id':'item_3_price_input'}))
+	item_3_price = forms.DecimalField(required=False, widget = forms.NumberInput(attrs={'id':'item_3_price_input'}))
 	item_3_job = forms.ModelChoiceField(queryset=Jobs.objects.all(), to_field_name="address", required=False, widget = forms.Select(attrs={'id':'item_3_job_input'})) # CHANGE LATER TO ONLY ACTIVE JOBS ON THIS
 	item_3_delivery_location = forms.ChoiceField(choices=(('shop', 'shop'),('site', 'site')), required=False, widget= forms.Select(attrs={'id':'item_3_delivery_location_input'})) # make it a radio widget, see docs on widgets
 	item_3_delivery_date = forms.DateField(required=False, widget = forms.SelectDateWidget(attrs={'id':'item_3_delivery_date_input'}))
@@ -56,7 +56,7 @@ class purchase_order_form(forms.Form):
 
 	item_4_description = forms.CharField(required=False, widget = forms.TextInput(attrs={'id':'item_4_description_input'}))
 	item_4_fullname = forms.CharField(required=False, widget = forms.TextInput(attrs={'id':'item_4_fullname_input'}))
-	item_4_price = forms.IntegerField(required=False, widget = forms.NumberInput(attrs={'id':'item_4_price_input'}))
+	item_4_price = forms.DecimalField(required=False, widget = forms.NumberInput(attrs={'id':'item_4_price_input'}))
 	item_4_job = forms.ModelChoiceField(queryset=Jobs.objects.all(), to_field_name="address", required=False, widget = forms.Select(attrs={'id':'item_4_job_input'})) # CHANGE LATER TO ONLY ACTIVE JOBS ON THIS
 	item_4_delivery_location = forms.ChoiceField(choices=(('shop', 'shop'),('site', 'site')), required=False, widget= forms.Select(attrs={'id':'item_4_delivery_location_input'})) # make it a radio widget, see docs on widgets
 	item_4_delivery_date = forms.DateField(required=False, widget = forms.SelectDateWidget(attrs={'id':'item_4_delivery_date_input'}))
@@ -64,7 +64,7 @@ class purchase_order_form(forms.Form):
 
 	item_5_description = forms.CharField(required=False, widget = forms.TextInput(attrs={'id':'item_5_description_input'}))
 	item_5_fullname = forms.CharField(required=False, widget = forms.TextInput(attrs={'id':'item_5_fullname_input'}))
-	item_5_price = forms.IntegerField(required=False, widget = forms.NumberInput(attrs={'id':'item_5_price_input'}))
+	item_5_price = forms.DecimalField(required=False, widget = forms.NumberInput(attrs={'id':'item_5_price_input'}))
 	item_5_job = forms.ModelChoiceField(queryset=Jobs.objects.all(), to_field_name="address", required=False, widget = forms.Select(attrs={'id':'item_5_job_input'})) # CHANGE LATER TO ONLY ACTIVE JOBS ON THIS
 	item_5_delivery_location = forms.ChoiceField(choices=(('shop', 'shop'),('site', 'site')), required=False, widget= forms.Select(attrs={'id':'item_5_delivery_location_input'})) # make it a radio widget, see docs on widgets
 	item_5_delivery_date = forms.DateField(required=False, widget = forms.SelectDateWidget(attrs={'id':'item_5_delivery_date_input'}))
@@ -72,7 +72,7 @@ class purchase_order_form(forms.Form):
 
 	item_6_description = forms.CharField(required=False, widget = forms.TextInput(attrs={'id':'item_6_description_input'}))
 	item_6_fullname = forms.CharField(required=False, widget = forms.TextInput(attrs={'id':'item_6_fullname_input'}))
-	item_6_price = forms.IntegerField(required=False, widget = forms.NumberInput(attrs={'id':'item_6_price_input'}))
+	item_6_price = forms.DecimalField(required=False, widget = forms.NumberInput(attrs={'id':'item_6_price_input'}))
 	item_6_job = forms.ModelChoiceField(queryset=Jobs.objects.all(), to_field_name="address", required=False, widget = forms.Select(attrs={'id':'item_6_job_input'})) # CHANGE LATER TO ONLY ACTIVE JOBS ON THIS
 	item_6_delivery_location = forms.ChoiceField(choices=(('shop', 'shop'),('site', 'site')), required=False, widget= forms.Select(attrs={'id':'item_6_delivery_location_input'})) # make it a radio widget, see docs on widgets
 	item_6_delivery_date = forms.DateField(required=False, widget = forms.SelectDateWidget(attrs={'id':'item_6_delivery_date_input'}))
@@ -80,7 +80,7 @@ class purchase_order_form(forms.Form):
 
 	item_7_description = forms.CharField(required=False, widget = forms.TextInput(attrs={'id':'item_7_description_input'}))
 	item_7_fullname = forms.CharField(required=False, widget = forms.TextInput(attrs={'id':'item_7_fullname_input'}))
-	item_7_price = forms.IntegerField(required=False, widget = forms.NumberInput(attrs={'id':'item_7_price_input'}))
+	item_7_price = forms.DecimalField(required=False, widget = forms.NumberInput(attrs={'id':'item_7_price_input'}))
 	item_7_job = forms.ModelChoiceField(queryset=Jobs.objects.all(), to_field_name="address", required=False, widget = forms.Select(attrs={'id':'item_7_job_input'})) # CHANGE LATER TO ONLY ACTIVE JOBS ON THIS
 	item_7_delivery_location = forms.ChoiceField(choices=(('shop', 'shop'),('site', 'site')), required=False, widget= forms.Select(attrs={'id':'item_7_delivery_location_input'})) # make it a radio widget, see docs on widgets
 	item_7_delivery_date = forms.DateField(required=False, widget = forms.SelectDateWidget(attrs={'id':'item_7_delivery_date_input'}))
@@ -88,7 +88,7 @@ class purchase_order_form(forms.Form):
 
 	item_8_description = forms.CharField(required=False, widget = forms.TextInput(attrs={'id':'item_8_description_input'}))
 	item_8_fullname = forms.CharField(required=False, widget = forms.TextInput(attrs={'id':'item_8_fullname_input'}))
-	item_8_price = forms.IntegerField(required=False, widget = forms.NumberInput(attrs={'id':'item_8_price_input'}))
+	item_8_price = forms.DecimalField(required=False, widget = forms.NumberInput(attrs={'id':'item_8_price_input'}))
 	item_8_job = forms.ModelChoiceField(queryset=Jobs.objects.all(), to_field_name="address", required=False, widget = forms.Select(attrs={'id':'item_8_job_input'})) # CHANGE LATER TO ONLY ACTIVE JOBS ON THIS
 	item_8_delivery_location = forms.ChoiceField(choices=(('shop', 'shop'),('site', 'site')), required=False, widget= forms.Select(attrs={'id':'item_8_delivery_location_input'})) # make it a radio widget, see docs on widgets
 	item_8_delivery_date = forms.DateField(required=False, widget = forms.SelectDateWidget(attrs={'id':'item_8_delivery_date_input'}))
@@ -96,7 +96,7 @@ class purchase_order_form(forms.Form):
 
 	item_9_description = forms.CharField(required=False, widget = forms.TextInput(attrs={'id':'item_9_description_input'}))
 	item_9_fullname = forms.CharField(required=False, widget = forms.TextInput(attrs={'id':'item_9_fullname_input'}))
-	item_9_price = forms.IntegerField(required=False, widget = forms.NumberInput(attrs={'id':'item_9_price_input'}))
+	item_9_price = forms.DecimalField(required=False, widget = forms.NumberInput(attrs={'id':'item_9_price_input'}))
 	item_9_job = forms.ModelChoiceField(queryset=Jobs.objects.all(), to_field_name="address", required=False, widget = forms.Select(attrs={'id':'item_9_job_input'})) # CHANGE LATER TO ONLY ACTIVE JOBS ON THIS
 	item_9_delivery_location = forms.ChoiceField(choices=(('shop', 'shop'),('site', 'site')), required=False, widget= forms.Select(attrs={'id':'item_9_delivery_location_input'})) # make it a radio widget, see docs on widgets
 	item_9_delivery_date = forms.DateField(required=False, widget = forms.SelectDateWidget(attrs={'id':'item_9_delivery_date_input'}))
@@ -104,7 +104,7 @@ class purchase_order_form(forms.Form):
 
 	item_10_description = forms.CharField(required=False, widget = forms.TextInput(attrs={'id':'item_10_description_input'}))
 	item_10_fullname = forms.CharField(required=False, widget = forms.TextInput(attrs={'id':'item_10_fullname_input'}))
-	item_10_price = forms.IntegerField(required=False, widget = forms.NumberInput(attrs={'id':'item_10_price_input'}))
+	item_10_price = forms.DecimalField(required=False, widget = forms.NumberInput(attrs={'id':'item_10_price_input'}))
 	item_10_job = forms.ModelChoiceField(queryset=Jobs.objects.all(), to_field_name="address", required=False, widget = forms.Select(attrs={'id':'item_10_job_input'})) # CHANGE LATER TO ONLY ACTIVE JOBS ON THIS
 	item_10_delivery_location = forms.ChoiceField(choices=(('shop', 'shop'),('site', 'site')), required=False, widget= forms.Select(attrs={'id':'item_10_delivery_location_input'})) # make it a radio widget, see docs on widgets
 	item_10_delivery_date = forms.DateField(required=False, widget = forms.SelectDateWidget(attrs={'id':'item_10_delivery_date_input'}))
